pyproject.toml
[project]
name = "composer-trade-py"
version = "0.0.0"
authors = [
    {name = "SolarWolf", email = "git@solarwol.xyz"},
]
description = "Python SDK for Composer.Trade API"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
keywords = ["trading", "finance", "api", "composer", "investment", "stocks", "algorithmic-trading"]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "hatch>=1.16.3",
    "twine>=6.2.0",
    "build>=1.4.0",
]

[project.urls]
Homepage = "https://github.com/composer-trade/composer-trade-py"
Documentation = "https://api.composer.trade/"
Repository = "https://github.com/composer-trade/composer-trade-py"
Issues = "https://github.com/composer-trade/composer-trade-py/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["composer"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "SIM"]
ignore = ["D100", "D104", "D107"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
